backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.4
sendgrid==6.11.0
websockets==12.0
apscheduler==3.10.4